requirements.txt
numpy==1.23.5
tensorflow-cpu==2.10.0
tensorflow-directml-plugin==0.4.0.dev230202
h5py==3.14.0
protobuf==3.19.6
keras==2.10.0
tensorboard==2.10.1
